requirements.txt
apache-airflow==1.10.10
boto3==1.12.40
botocore==1.15.40
docutils==0.15.2
jmespath==0.9.5
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
